+++ spring-toby/spring-project/src/main/java/com/javabom/toby/user/connectionmaker/ConnectionMaker.java
@@ -1,0 +1,5 @@
+package com.javabom.toby.user.connectionmaker;
+
+public interface ConnectionMaker {
+    void makeConnection();
+}
