+++ spring-toby/spring-project/src/main/java/com/javabom/toby/user/dao/DaoFactory.java
@@ -1,7 +1,7 @@
-package com.javabom.toby.userdao;
+package com.javabom.toby.user.dao;
 
-import com.javabom.toby.userdao.connectionmaker.ConnectionMaker;
-import com.javabom.toby.userdao.connectionmaker.LocalDBConnectionMaker;
+import com.javabom.toby.user.connectionmaker.ConnectionMaker;
+import com.javabom.toby.user.connectionmaker.LocalDBConnectionMaker;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.jdbc.datasource.SimpleDriverDataSource;
