+++ spring-toby/spring-project/src/main/java/com/javabom/toby/user/dao/StatementStrategy.java
@@ -1,4 +1,4 @@
-package com.javabom.toby.userdao;
+package com.javabom.toby.user.dao;
 
 import java.sql.Connection;
 import java.sql.PreparedStatement;
