--- spring-toby/spring-project/src/main/java/com/javabom/toby/userdao/connectionmaker/ConnectionMaker.java
@@ -1,7 +1,0 @@
-package com.javabom.toby.userdao.connectionmaker;
-
-import java.sql.Connection;
-
-public interface ConnectionMaker {
-    void makeConnection();
-}
